141. version 2.0
I use set to record the node we have found. If we find a node twice,
then there is a cycle. It works but it's slow. Because the time
complexity is O(nlogn), and space complexity is O(n). Meanwhile, the
fast and slow pointers are O(n) and O(1)